Necesary changes after changing GridInfo and CellInfo from int to long

+++ src/edu/umn/cs/WriteRectFile.java
@@ -55,8 +55,8 @@ public class WriteRectFile {
 	    
 	    FSDataOutputStream os = fs.create(getCellFilePath(column, row), gridInfo);
 	    cellStreams[column][row] = ps = new PrintStream(os);
-      int xCell = column * gridInfo.cellWidth + gridInfo.xOrigin;
-      int yCell = row * gridInfo.cellHeight + gridInfo.yOrigin;
+      long xCell = column * gridInfo.cellWidth + gridInfo.xOrigin;
+      long yCell = row * gridInfo.cellHeight + gridInfo.yOrigin;
       System.out.println("Setting next block at "+xCell+","+ yCell);
 	    ((DFSOutputStream)os.getWrappedStream()).setNextBlockCell(xCell, yCell);
 	  }
